auth in header

+++ src/client/main.rb
@@ -144,7 +144,7 @@ end
 def perform_post(params, content_type='application/json')
   post_params = {}
   post_params[:api_vars] = params
-   add_access(post_params)
+ #  add_access(post_params)
   #  STDERR.puts  @route
 
   rest_post(@route,post_params, content_type)
@@ -231,7 +231,7 @@ def rest_get(path,params=nil)
     retry_count = 0
     # STDERR.puts('Get Path:' + path.to_s + ' Params:' + params.to_s)
     params = add_access(params)
-    r = RestClient.get(@base_url + path, params)
+    r = RestClient.get(@base_url + path, params, {  :access_token => load_token})
 
     return r
   rescue RestClient::ExceptionWithResponse => e
@@ -263,9 +263,10 @@ def rest_post(path, params, content_type )
 
   begin
      
-    params = add_access(params)
+   # params = add_access(params)
     #STDERR.puts('Post Path:' + path.to_s + ' Params:' + params.to_s)
-    r = RestClient.post(@base_url + path, params.to_json, { :content_type => content_type, :access_token => load_token} )
+    params = params.to_json if content_type == 'application/json'
+    r = RestClient.post(@base_url + path, params, { :content_type => content_type, :access_token => load_token} )
 #    unless content_type.nil?
 #      #  STDERR.puts  'ct ' + content_type
 #      #   r = RestClient.post(@base_url + path, params[:api_vars][:data], :content_type => content_type )
@@ -287,10 +288,10 @@ def rest_post(path, params, content_type )
 end
 
 def rest_delete(path, params=nil)
-  params = add_access(params)
+ # params = add_access(params)
   begin
     #STDERR.puts('Post Path:' + path.to_s + ' Params:' + params.to_s)
-    r = RestClient.delete(@base_url + path, params)
+    r = RestClient.delete(@base_url + path, params, {  :access_token => load_token} )
     write_response(r)
     exit
   rescue RestClient::ExceptionWithResponse => e
